shop_api/requirements.txt
amqp==5.3.1
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.4.2
lxml==5.3.0
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg==3.1.12
psycopg2-binary==2.9.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
XlsxWriter==3.2.0
